+++ src/background/permissionPrompt.ts
